+++ examples/cmdapp/cmd/app/model.go
@@ -1,0 +1,3 @@
+package main
+
+const appSubtitle = "Child entry package calling internal GOX packages"